tests: initial setup for Candlepin-based tests
Integration tests for the rhc role will definitely need a Candlepin
server for registration.

As a way to easily test scenarions, add an helper playbook that either
deploys Candlepin using the upstream Ansible role, or assumes an
external Candlepin server is available anywhere; describe this in a
README.md file.

The approach is taken from the current way the Ansible role is used in
the Cockpit CI to generate an image with Candlepin installed as a
container.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>