Implement basic registration/unregistration
Add a very simple/minimal task to call the "redhat_subscription" module
to register & unregister.

Extend the API with basic parameters:
- add "rhc_state" for the registration state of the system
- add "rhc_auth" for authentication parameters
- add "rhc_organization" for the user organization, in case it must be
  specified
- add "rhc_server" to handle the connection options to the registration
  server

All the parameters are currently directly passed to the
"redhat_subscription" module; there is no additional validation for
them -- it is planned for later steps.

Add a very simple integration test that tries to first register and then
unregister, doing each step twice to verify the idempotency of the role
on status changes.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>