tests: always disable cert validation when checking Candlepin
Properly checking that the SSL connection to Candlepin works requires
also the CA of that Candlepin instance, which is generally not in the
system CA store (but in /etc/rhsm/ca for subscription-manager).

Since this is only an existance check, disable the certificate
validation. In case of SSL problems, the actual subscription-manager
invocation will fail anyway.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>