tests: enable content from deployed Candlepin
When deploy Candlepin in the test system, tell it to also configure the
repositories with test content for the test data (i.e. "-r" for the
deploy script). This way it will be possible to simulate fetching actual
content from it.

Fetch the GPG key for the repositories, so RPM can install the test
packages.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>