Implement rhc_baseurl
Add a parameter for specifying the base URL for the content received
from the subscription server.

This is set as separate parameter on its own, because the
redhat_subscription module both configures it and passes it as argument
to "subscription-manager register --baseurl"; the problem with --baseurl
is that subscription-manager parses it and then assumes https as
protocol, breaking the original URL set by the user.

Extend the test setup so it is possible to specify it.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>