Ignore candlepin bits
During the tests we can checkout an external repository that contains
Ansible roles. Since that is external code not related to this
repository and that has its own style/policies/etc, make sure to ignore
it locally so it is less in the way when running checkers/linters.

This needs an own config section in tox.ini to make sure that the
yamllint check done by tox-lsr will use the .yamllint.yml file in this
repository instead of its own copy. Thanks to Richard Megginson for the
hint!

Signed-off-by: Pino Toscano <ptoscano@redhat.com>