Workaround redhat_subscription requiring creds even when unneeded (#29)
The 'redhat_subscription' module currently requires parameters for
credentials in case state=present. This applies also in case the system
is already registered (and force_register=false), being then effectively
unused. This is proposed as change to 'redhat_subscription' as:
https://github.com/ansible-collections/community.general/pull/5664

In the meanwhile, workaround this situation locally in the role:
- add a fake credential string as private variable
- pass the fake credential string to both 'username' and 'password' in
  case rhc_state=present and the system is already registered

This way, all the users of the role do not need to specify credentials
when not needed; this reflects in the tests, from which it is possible
to drop the extra credentials.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>

Signed-off-by: Pino Toscano <ptoscano@redhat.com>