Implement rhc_environments
Add a parameter for configuring the environments to register to when
connecting (and only in that case, due to different permissions needed
to change the environments in already connected systems).

Make sure to test this:
- extend the test setup with variables related to environments
- add an helper script to get the list of set environments from
  subscription-manager
- improve the Candlepin deployment by adding a couple of test
  environments
- add a new test to verify that setting environments on connection works

Since enabling environments makes the environment selection mandatory,
pass the configured environments to all the invocations of "rhc" that
actually do connect successfully.