tests: improve assertions for negative tests
In case we expect a task to fail, we check in the "rescue" part of a
"block" that something actually failed; since the "block" has a "fail"
task to force a failure in case the task did not fail, we want to check
that the reason we get into "rescue" is not the "forced failure".

Hence, tighten the asserts to ignore the forced failure task.
Suggested by Richard Megginson in
https://github.com/linux-system-roles/rhc/pull/48

Signed-off-by: Pino Toscano <ptoscano@redhat.com>