Implemented "rhc_insights.tags" parameter
Add a new parameter to configure insights tags that will be added
to the system record in HBI.

The syntax of this parameter is a dictionary, and the value for
each key can be any data type.

The role will try to configure the tags if this parameter is present.
If it is not present the tags will remain in the same status.

Test added that verifies:
- Add and delete tags
- Remove all tags
- Do not modify tags if the parameter is missing.

Signed-off-by: Alba Hita Catala <ahitacat@redhat.com>