tests: do not deploy Candlepin in tests_release
A local Candlepin deployment has no concept of "releases", so deploying
it to unconditionally end the test is suboptimal.

Make use of the new helper tasks for the tests:
- only setup the test data first, so we get the configured release
  version for the test
- keep skipping the test if there is no configured release: this way,
  even an external Candlepin configuration may still skip the test
- ensure that the configured Candlepin works: setup_candlepin.yml does
  it automatically, so we need to manually use check_candlepin.yml now

The only behaviour change is that this test will end much faster when
using the local Candlepin deployment (i.e. the default setup).

Signed-off-by: Pino Toscano <ptoscano@redhat.com>