tests: pin Candlepin version to latest before Java 17
The current way of deploying Candlepin uses an Ansible role that is no
more used/supported by the Candlepin project, and thus it is rotting;
the recent switch of Candlepin from Java 11 to Java 17 broke with the
unmaintained role.

Thus, as very short-term workaround, pin the version of Candlepin to
deploy to the commit before the switch to Java 17.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>