tests: switch to my unofficial Candlepin container image
The current way to test registration is to deploy Candlepin from sources
in the managed node; while this initially worked fine, it started to
grow a number of drawbacks:
- building Candlepin from sources (including importing the test data for
  it) takes a lot of time, at least 5-6 minutes even on fast systems
- the Candlepin project changes the supported OSes over time, and thus
  it is not possible to test all the supported OSes of the role
- the currently used Ansible role is no more used/supported by the
  Candlepin project, and thus it is rotting
In the end, we really don't need to build Candlepin from sources, but
rather only use an already deployed Candlepin.

The Candlepin people are working on an official Candlepin container
image usable also outside of their CI system; since this process will
take some time, I built an unofficial container image of Candlepin:
  https://github.com/ptoscano/candlepin-container-unofficial
It contains everything (DB + Candlepin + test data) in a single image.
While this is unofficial, it is not worse than the previously used way
of deploying Candlepin, and it has the advantage of being stable (i.e.
existing tags should still work fine regardless of upstream changes).

Hence, switch to this image for now:
- install podman to manage the container
- start the container, waiting for Candlepin to start only when needed
  (delaying it to right before the tasks that need to use Candlepin)
- copy the certificates from the running container
- remove everything else related to the Candlepin deployment
- use "candlepin.local" as hostname for Candlepin, as it is the hostname
  used during the container build, and thus the self-signed certificates
  of Candlepin use that hostname; hence, set it as hostname for the
  container, and add the entry for it in /etc/hosts
Also, since the OS needed for running Candlepin must not be the same OS
of the managed node, the OS limitation is dropped.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>