fix: correctly set release using new behavior of rhsm_release
The way the new version of rhsm_release works is this:

- if the `release` parameter is present, set the release to the given
  value.
- If there is no `release` parameter, unset the release.

If the user does not set `rhc_release`, or if the user sets it to `omit`
then it will be `null`, and calling the rhsm_release module will be
skipped.
See https://github.com/ansible-collections/community.general/pull/6401
for details.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>