refactor: drop fake credential workaround
The redhat_subscription module used to always require credentials when
"state=present"; since we used the module also to do basic configuration
of subscription-manager, this required us to pass fake credentials
(unused) when the system was already registered. This behaviour was
changed with [1] (backported to community.general 6.x with [2]), and
available since community.general 6.5.0.

Hence, drop the concept of fake credential, and pass "username" and
"password" only when specified; redhat_subscription still properly
checks that only one credential type is specified.

[1] https://github.com/ansible-collections/community.general/pull/5664
[2] https://github.com/ansible-collections/community.general/pull/6222

Signed-off-by: Pino Toscano <ptoscano@redhat.com>