feat: add ansible host to insights configuration
Adds the ansible host name option for insights-client
into the rhc role. This adds support for setting an
ansible host name in HBI during insights-client
registration, updating the current ansible host name
in HBI if the system is already registered, and
resetting the ansible host name if an absent value is provided.

Signed-off-by: Jason Jerome <jajerome@redhat.com>