fix: Ignore ansible_host: ""
rhc.insights.ansible_host set to an empty string causes:

- the Ansible host name to be reset in the Inventory by
`insights-client --ansible_host=`;
- a line `ansible_host=` to be put in the config file.

The former is a job of `{state: absent}`. An empty string should not
cause such a desctructive operation. The latter is ignored by the
Client and is equivatent to that line missing.

Consistently with a similar condition of the display_name parameter,
an empty string ansible_host is treated as undefined. It’s the same
behavior as with a null value.