fix: drop usage of "auto_attach" of the "redhat_subscription" module
"auto_attach" is supposed to automatically attach subscriptions for
accounts in entitlement mode. in Simple Content Access mode, this option
becomes a no-op in subscription-manager.

The usage of this option was originally added to help testing regardless
of the content access mode, even if the role supports only SCA. Since
the entitlement mode has been officially deprecated for a couple of
years already, and it is going away soon, then stop using the
"auto_attach" option of the "redhat_subscription" module.

There is no behaviour change.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>