tests: factorize & fix URL for candlepin owner
The Candlepin URL used to query the owner (= organization) used in the
tests when using the self-deployed Candlepin was wrong, using the
username rather than the organization name. It works fine because the
username and the organization name are currently the same.

Hence, to avoid issues in the future, properly use the organization name
instead. To simplify the maintenance of that code, create a new helper
fact with the Candlepin URL for the current owner.

There is no behaviour change in the tests.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>