tests: run podman directly instead of using containers.podman
Using the containers.podman collection to handle podman is a nice idea,
however what that collection supports now is becoming more strict than
what this role supports. Case in point: containers.podman now requires
Python 3.6+ for the managed nodes, which breaks the testing of EL 7.

This means that sadly the setup of a self-deployed Candlepin needs to
run podman commands manually; use the proper parameters for "podman run"
depending on the OS.

The gymnastics of checking the return code & stderr of "podman stop"
are needed because the "--ignore" option (which would be ideal) was
introduced in podman 1.7.0, and EL 7 has podman 1.6.4.

There is no behaviour change.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>