tests: stop adding pools to the activation key
The Candlepin account used for tests uses SCA, and thus any activation
key for it can already access to the whole content of the organization
of the account. The addition of the pools to the activation key is thus
redundant, and upcoming versions of Candlepin will not allow this
behaviour anymore.

Hence, stop adding pools to the activation key, which is no more needed.

There is no behaviour change.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>